+++ .github/workflows/ci.yml
@@ -17,8 +17,8 @@ jobs:
         os: [ubuntu-latest]
 
     steps:
-      - uses: "actions/checkout@v2"
-      - uses: "actions/setup-python@v1"
+      - uses: "actions/checkout@v3"
+      - uses: "actions/setup-python@v4"
         with:
             python-version: "${{ matrix.python-version }}"
       - name: Run CI tests
